requirements.txt
openai==1.12.0
Flask==3.0.2
python-dotenv==1.0.1
